feat: add conventional-changelog utility
