+++ package.json
@@ -8,7 +8,7 @@
     "index.js"
   ],
   "scripts": {
-    "test": "c8 mocha test.js",
+    "test": "c8 mocha ./test/*.js",
     "test:snap": "CHAI_JEST_SNAPSHOT_UPDATE_ALL=true npm test",
     "posttest": "standard",
     "fix": "standard --fix",
@@ -37,5 +37,8 @@
     "standard": "^16.0.3",
     "unist-util-inspect": "^6.0.1",
     "yargs": "^16.2.0"
+  },
+  "dependencies": {
+    "unist-util-visit": "^2.0.3"
   }
 }
